Correct H(div) pyramid, interior Family III
Was missing the factor of n (ijmax) from the derivative of mu^zeta_0, and was using grad mu_1^zeta instead of mu_0^zeta; see Fuentes et al. p. 454 for the definition.  (ESEAS still swaps arguments for phi_quad relative to the definition, but this just leads to a reordering of the basis, no changes to their values.)